+++ era_customer_success/models/__init__.py
@@ -14,3 +14,4 @@ from . import voip_call
 from . import survey_user_input
 from . import ir_attachment
 from . import cs_weekly_suggestion
+from . import cs_success_profile